feat: Create db module to process db configuration
